Refactor spacemap
